Changed Scraper result from Metric to the scraper own result representation.
Previously, Scraper returned Metric as the result of Scrape method. This
resulted in a redundant dependency between scraper and metric packages. By
introducing the Result struct for the Scrape method the dependency between
packages was broken down. Scrape can now have an independent model which is
supposed to be assembled into a required model (like metric) on a user side
(config package).