Refactored scraper.
A scraper produces one result on each scrape function invocation.
Introduced a producer which is a wrapper over scraper that runs it in an
infinite loop to have a stream of scrape results.
Introduced a producer manager which exposes methods to work with producers in
memory.
Made scraper private as the only entity supposed to be exported from the package
 is InMemoryManager.